requirements.txt
beautifulsoup4==4.11.1
firebase==3.0.1
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-firestore==2.7.2
googleapis-common-protos==1.56.4
selenium==4.7.2
grpcio==1.50.0
grpcio-status==1.50.0
lxml==4.9.2
pytest==7.2.0
mock-firestore==0.11.0